[LCR-113] feat: 삭제 기능 추가
